Add feature to remove background
If the user specifies the backgroundcolour color as "none", then generate the document without adding a background rectangle. This allows the SVG outputs to be used in applications that either require a transparent background or where the rectangle itself will be printed, e.g., on pen plotters and laser cutters.

Also fix a typo in UI help.